backend-django/requirements.txt
Django==4.2.11
djangorestframework==3.15.2
drf-yasg==1.21.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-cors-headers==4.9.0
haystack-ai==2.23.0
